display (sms) header info in netmon1 screen
src:dst:sap
