+++ applet/src/radiostate.h
@@ -6,6 +6,8 @@
 #ifndef RADIOSTATE_H
 #define RADIOSTATE_H
 
+#include "etsi.h"
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -15,10 +17,16 @@ extern int rst_src ;
 extern int rst_dst ;
 extern int rst_mycall ;
     
+extern int rst_hdr_sap ;
+extern int rst_hdr_src ;
+extern int rst_hdr_dst ;
+
 void rst_voice_lc_header(int src, int dst);
 
 void rst_term_with_lc( int src, int dst );
 
+void rst_data_header(data_t *data);
+        
 void rst_signal_other_call();
 void rst_signal_my_call();
